pyproject.toml
[project]
name = "book-reviews"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.4",
    "bleach>=6.2.0",
    "flask>=3.1.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "markdown>=3.8.2",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "ruff>=0.15.4",
    "sqlalchemy>=2.0.48",
]


[tool.ruff]
# Set the maximum line length to 79.
line-length = 79

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter
extend-select = ["E501"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"site/testing/*.py" = ["ALL"]
"test_*.py" = ["ALL"]
